pyproject.toml
[project]
name = "agenticai-base-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.0.0",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "mcp>=1.12.3",
    "ollama>=0.5.1",
    "openai>=1.98.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "semantic-kernel>=1.32.2",
]
